🎨 Add AppLdapEntry component update LdapEntry
